--- timer.py
@@ -1,10 +1,0 @@
-from threading import Timer
-
-
-class RepeatTimer(Timer):
-
-    def run(self):
-        while not self.finished.wait(self.interval):
-            self.function(*self.args, **self.kwargs)
-
-
